Added start of collision detection

+++ src/Game/Shapes/Shape.java
@@ -5,11 +5,11 @@ import javafx.scene.layout.Pane;
 import java.util.List;
 
 public interface Shape {
-    int[] VALID_X_COORDINATES = {285, 300, 315, 330};
-    int[] VALID_Y_COORDINATES = {75, 90};
-    int[] VALID_NEXT_X_COORDINATES = {420, 435, 450, 465};
-    int[] VALID_NEXT_Y_COORDINATES = {210, 225};
-    int[] VALID_COORDINATE_MODIFIERS = {15, 30};
+    double[] VALID_X_COORDINATES = {285, 300, 315, 330};
+    double[] VALID_Y_COORDINATES = {75, 90};
+    double[] VALID_NEXT_X_COORDINATES = {420, 435, 450, 465};
+    double[] VALID_NEXT_Y_COORDINATES = {210, 225};
+    double[] VALID_COORDINATE_MODIFIERS = {15, 30};
     int LEVEL_IMAGE_LOOP = 10;
 
     int getColumn();
